start to use preacte for client side componenet
