+++ packages/tc-ui-relationships/browser.jsx
@@ -1,0 +1,3 @@
+const { RelationshipPicker } = require('./edit');
+
+module.exports = { RelationshipPicker };
